+++ README.md
@@ -38,4 +38,4 @@ npm install
 npm run start
 ```
 
-SLOC: 304
+SLOC: 317
