+++ client/next.config.js
