+++ client/src/app/api/auth/[...nextauth]/route.ts
@@ -1,6 +1,6 @@
 import NextAuth from 'next-auth'
 
-import { authOptions } from '@/lib/auth'
+import { authOptions } from '../../../../lib/auth'
 
 const handler = NextAuth(authOptions)
