+++ client/src/app/favicon.ico
