+++ client/src/app/page.tsx
@@ -1,6 +1,6 @@
 import { getServerSession } from 'next-auth/next';
 
-import { authOptions } from '@/lib/auth';
+import { authOptions } from '../lib/auth';
 import { redirect } from 'next/navigation';
 
 export default async function Home() {
