+++ client/src/components/MyNavbar.tsx
@@ -1,6 +1,6 @@
 import { getServerSession } from 'next-auth/next';
 
-import { authOptions } from '@/lib/auth';
+import { authOptions } from '../lib/auth';
 
 export const MyNavbar = async () => {
   const session = await getServerSession(authOptions);
