+++ client/src/lib/auth.ts
@@ -1,7 +1,7 @@
 import GoogleProvider from 'next-auth/providers/google';
 import { PrismaAdapter } from '@next-auth/prisma-adapter';
 
-import { prisma } from '@/lib/db';
+import { prisma } from './db';
 
 export const authOptions = {
   callbacks: {
