+++ package.json
@@ -1,39 +1,14 @@
 {
   "name": "bookkss",
-  "version": "0.1.0",
-  "private": true,
+  "version": "1.0.0",
+  "description": "Bookkss provides a platform for users to write and share book reviews.",
+  "main": "index.js",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "postinstall": "prisma generate",
-    "server": "nodemon src/server/index.js"
+    "client": "cd client && npm run dev",
+    "server": "cd server && npm run start",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "dependencies": {
-    "@headlessui/react": "^1.7.15",
-    "@heroicons/react": "^2.0.18",
-    "@next-auth/prisma-adapter": "^1.0.5",
-    "@prisma/client": "^5.0.0",
-    "@types/react": "18.2.15",
-    "@types/react-dom": "18.2.7",
-    "autoprefixer": "10.4.14",
-    "axios": "^1.4.0",
-    "eslint": "8.45.0",
-    "eslint-config-next": "13.4.10",
-    "express": "^4.18.2",
-    "next": "13.4.10",
-    "next-auth": "4.22.1",
-    "postcss": "8.4.26",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-icons": "^4.10.1",
-    "tailwindcss": "3.3.3",
-    "typescript": "5.1.6"
-  },
-  "devDependencies": {
-    "@types/node": "^20.4.2",
-    "prettier": "^3.0.0",
-    "prettier-plugin-tailwindcss": "^0.4.1"
-  }
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
 }
